test: split Anthropic OAuth persistence coverage
