+++ tests2/core/anthropic-oauth-persistence.test.ts
@@ -3,25 +3,11 @@ guardProcessEnv();
 
 import assert from "node:assert/strict";
 import { randomUUID } from "node:crypto";
-import {
-	existsSync,
-	mkdirSync,
-	mkdtempSync,
-	readFileSync,
-	readdirSync,
-	realpathSync,
-	rmdirSync,
-	rmSync,
-	statSync,
-	symlinkSync,
-	utimesSync,
-	writeFileSync,
-} from "node:fs";
+import { mkdirSync, mkdtempSync, readFileSync, rmSync } from "node:fs";
 import { tmpdir } from "node:os";
 import path from "node:path";
 import { Worker } from "node:worker_threads";
-import type { Credential, OAuthCredential } from "@earendil-works/pi-ai";
-import { ModelRuntime } from "@earendil-works/pi-coding-agent";
+import type { OAuthCredential } from "@earendil-works/pi-ai";
 import { afterAll, afterEach, beforeEach, describe, it, vi } from "vitest";
 import { resetAgentDirStateForTests } from "../../src/server/bobbit-dir.js";
 import { AtomicCredentialStore } from "../../src/server/auth/credential-store.js";
@@ -35,7 +21,6 @@ resetAgentDirStateForTests();
 
 const {
 	getOAuthCredentialStore,
-	getOAuthModels,
 	oauthStatus,
 	refreshGoogleOAuthToken,
 	refreshOAuthToken,
@@ -43,24 +28,6 @@ const {
 } = await import("../../src/server/auth/oauth.js");
 
 const realFetch = globalThis.fetch;
-
-/**
- * A credential-free stand-in for Pi's proper-lockfile protocol. Pi resolves
- * the auth file's real path, then atomically creates and heartbeats this empty
- * sibling directory. Keeping the fixture here avoids requiring a root-level
- * proper-lockfile installation in the no-install unit-test gate.
- */
-function acquirePiCompatibleExternalLock(file: string): () => Promise<void> {
-	const lockPath = `${realpathSync(file)}.lock`;
-	mkdirSync(lockPath, { mode: 0o700 });
-	const heartbeat = setInterval(() => {
-		utimesSync(lockPath, new Date(), new Date());
-	}, 1_000);
-	return async () => {
-		clearInterval(heartbeat);
-		rmdirSync(lockPath);
-	};
-}
 
 function credential(expires = Date.now() + 60 * 60 * 1000): OAuthCredential {
 	return {
@@ -118,227 +85,6 @@ afterEach(() => {
 afterAll(() => {
 	stopFlowCleanup();
 	rmSync(tmp, { recursive: true, force: true });
-});
-
-describe("AtomicCredentialStore", () => {
-	it("serializes independent store instances and preserves mixed providers", async () => {
-		const stores = [
-			new AtomicCredentialStore(authPath),
-			new AtomicCredentialStore(authPath),
-			new AtomicCredentialStore(authPath),
-		];
-		const apiKey: Credential = { type: "api_key", key: randomUUID() };
-		await stores[0].modify("custom-provider", async () => apiKey);
-
-		const anthropic = credential();
-		const google = credential();
-		const codex = credential();
-		await Promise.all([
-			stores[0].modify("anthropic", async () => {
-				await new Promise((resolve) => setTimeout(resolve, 20));
-				return anthropic;
-			}),
-			stores[1].modify("google-gemini-cli", async () => google),
-			stores[2].modify("openai-codex", async () => codex),
-		]);
-
-		const document = readDocument();
-		assert.deepEqual(Object.keys(document).sort(), [
-			"anthropic",
-			"custom-provider",
-			"google-gemini-cli",
-			"openai-codex",
-		]);
-		sameSecret(document.anthropic.access, anthropic.access);
-		sameSecret(document["google-gemini-cli"].refresh, google.refresh);
-		sameSecret(document["openai-codex"].access, codex.access);
-		sameSecret(document["custom-provider"].key, apiKey.key);
-	});
-
-	it("keeps unrelated fresh reads available while a locked provider callback is stalled", async () => {
-		const store = new AtomicCredentialStore(authPath);
-		const anthropic = credential();
-		const google = credential();
-		await store.modify("anthropic", async () => anthropic);
-		await store.modify("google-gemini-cli", async () => google);
-		let release!: () => void;
-		const stalled = new Promise<void>((resolve) => { release = resolve; });
-		let markStarted!: () => void;
-		const started = new Promise<void>((resolve) => { markStarted = resolve; });
-		const mutation = store.modify("anthropic", async () => {
-			markStarted();
-			await stalled;
-			return credential();
-		});
-		await started;
-
-		const unrelatedRead = store.read("google-gemini-cli");
-		const completedWhileStalled = await Promise.race([
-			unrelatedRead.then(() => true),
-			new Promise<boolean>((resolve) => setTimeout(() => resolve(false), 150)),
-		]);
-		assert.equal(completedWhileStalled, true, "unrelated provider access must not wait for provider I/O");
-		release();
-		await mutation;
-	});
-
-	it("coordinates mutations with Pi's realpath lock while fresh reads remain asynchronous", async (context) => {
-		const realDir = path.join(tmp, `real-agent-${randomUUID()}`);
-		const aliasDir = path.join(tmp, `alias-agent-${randomUUID()}`);
-		mkdirSync(realDir, { recursive: true });
-		try {
-			symlinkSync(realDir, aliasDir, process.platform === "win32" ? "junction" : "dir");
-		} catch {
-			context.skip();
-			return;
-		}
-		const realAuthPath = path.join(realDir, "auth.json");
-		const aliasAuthPath = path.join(aliasDir, "auth.json");
-		await new AtomicCredentialStore(realAuthPath).modify("anthropic", async () => credential());
-		const releaseExternal = acquirePiCompatibleExternalLock(realAuthPath);
-		const aliasedStore = new AtomicCredentialStore(aliasAuthPath);
-		let mutationEntered = false;
-		let released = false;
-		try {
-			const aliasedMutation = aliasedStore.modify("anthropic", async () => {
-				mutationEntered = true;
-				return undefined;
-			});
-			const result = await Promise.race([
-				aliasedStore.read("anthropic"),
-				new Promise<never>((_, reject) => setTimeout(() => reject(new Error("fresh read waited for writer lock")), 150)),
-			]);
-			assert.equal(result?.type, "oauth");
-			await new Promise((resolve) => setTimeout(resolve, 50));
-			assert.equal(mutationEntered, false, "canonical mutation must contend on Pi's realpath lock");
-			await releaseExternal();
-			released = true;
-			await aliasedMutation;
-			assert.equal(mutationEntered, true);
-		} finally {
-			if (!released) await releaseExternal().catch(() => {});
-		}
-	});
-
-	it("holds the canonical lock across refresh callbacks so one rotating token is spent", async () => {
-		const first = new AtomicCredentialStore(authPath);
-		const second = new AtomicCredentialStore(authPath);
-		await first.modify("anthropic", async () => credential(Date.now() - 60_000));
-		const rotated = credential();
-		let refreshCallbacks = 0;
-		const rotate = (store: AtomicCredentialStore) => store.modify("anthropic", async (current) => {
-			if (current?.type !== "oauth" || Date.now() < current.expires) return undefined;
-			refreshCallbacks += 1;
-			await new Promise((resolve) => setTimeout(resolve, 50));
-			return rotated;
-		});
-
-		const [firstResult, secondResult] = await Promise.all([rotate(first), rotate(second)]);
-		assert.equal(refreshCallbacks, 1, "the second process must re-check the winner before refreshing");
-		const stored = await first.read("anthropic");
-		assert.equal(stored?.type, "oauth");
-		assert.equal(firstResult?.type, "oauth");
-		assert.equal(secondResult?.type, "oauth");
-		if (stored?.type === "oauth" && firstResult?.type === "oauth" && secondResult?.type === "oauth") {
-			sameSecret(firstResult.access, stored.access);
-			sameSecret(secondResult.access, stored.access);
-			sameSecret(stored.access, rotated.access);
-		}
-	});
-
-	it("serializes refresh with Pi's default external file-backed CredentialStore", async () => {
-		const expired = credential(Date.now() - 60_000);
-		const google = credential();
-		const codex = credential();
-		const store = new AtomicCredentialStore(authPath);
-		await store.modify("anthropic", async () => expired);
-		await store.modify("google-gemini-cli", async () => google);
-		await store.modify("openai-codex", async () => codex);
-
-		// Omitting `credentials` exercises Pi's own AuthStorage rather than a
-		// second Bobbit adapter instance. It snapshots the expired row before the
-		// Bobbit runtime starts refreshing, exactly like a separate Pi process.
-		const externalPi = await ModelRuntime.create({
-			authPath,
-			modelsPath: null,
-			allowModelNetwork: false,
-		});
-		const bobbitPi = getOAuthModels();
-		const rotatedAccess = randomUUID();
-		const rotatedRefresh = randomUUID();
-		let networkRequests = 0;
-		let markStarted!: () => void;
-		const started = new Promise<void>((resolve) => { markStarted = resolve; });
-		let releaseNetwork!: () => void;
-		const stalledNetwork = new Promise<void>((resolve) => { releaseNetwork = resolve; });
-		globalThis.fetch = (async () => {
-			networkRequests += 1;
-			markStarted();
-			await stalledNetwork;
-			return new Response(JSON.stringify({
-				access_token: rotatedAccess,
-				refresh_token: rotatedRefresh,
-				expires_in: 3600,
-			}), { status: 200, headers: { "Content-Type": "application/json" } });
-		}) as typeof fetch;
-
-		const bobbitRefresh = bobbitPi.getAuth("anthropic");
-		await started;
-		const externalRefresh = externalPi.getAuth("anthropic");
-		await new Promise((resolve) => setTimeout(resolve, 100));
-		const requestsBeforeRelease = networkRequests;
-		releaseNetwork();
-		const [bobbitResolution, externalResolution] = await Promise.all([bobbitRefresh, externalRefresh]);
-		assert.equal(requestsBeforeRelease, 1, "external Pi must wait and reuse Bobbit's persisted rotation");
-		assert.equal(typeof bobbitResolution?.auth.apiKey, "string");
-		assert.equal(typeof externalResolution?.auth.apiKey, "string");
-		assert.equal(bobbitResolution?.auth.apiKey === externalResolution?.auth.apiKey, true);
-		const document = readDocument();
-		sameSecret(document.anthropic.access, rotatedAccess);
-		sameSecret(document.anthropic.refresh, rotatedRefresh);
-		sameSecret(document["google-gemini-cli"].access, google.access, "Anthropic refresh must preserve Google");
-		sameSecret(document["openai-codex"].access, codex.access, "Anthropic refresh must preserve Codex");
-	});
-
-	it("freshly reads external changes instead of retaining a credential snapshot", async () => {
-		const first = new AtomicCredentialStore(authPath);
-		const second = new AtomicCredentialStore(authPath);
-		const initial = credential();
-		const rotated = credential();
-		await first.modify("anthropic", async () => initial);
-		const firstRead = await first.read("anthropic");
-		assert.equal(firstRead?.type, "oauth");
-		sameSecret(firstRead?.type === "oauth" ? firstRead.access : undefined, initial.access);
-
-		await second.modify("anthropic", async () => rotated);
-		const refreshedRead = await first.read("anthropic");
-		assert.equal(refreshedRead?.type, "oauth");
-		sameSecret(refreshedRead?.type === "oauth" ? refreshedRead.access : undefined, rotated.access);
-	});
-
-	it("fails closed on malformed external writes without replacing their bytes", async () => {
-		const malformed = `{${randomUUID()}`;
-		writeFileSync(authPath, malformed, "utf8");
-		const store = new AtomicCredentialStore(authPath);
-		await assert.rejects(
-			store.modify("anthropic", async () => credential()),
-			/invalid JSON|JSON object/i,
-		);
-		assert.equal(readFileSync(authPath, "utf8") === malformed, true);
-	});
-
-	it("uses restrictive permissions and removes same-directory temporary files", async () => {
-		const store = new AtomicCredentialStore(authPath);
-		await store.modify("anthropic", async () => credential());
-		assert.equal(existsSync(authPath), true);
-		if (process.platform !== "win32") {
-			assert.equal(statSync(authPath).mode & 0o777, 0o600);
-		}
-		assert.deepEqual(
-			readdirSync(agentDir).filter((name) => name.includes(".bobbit-") && name.endsWith(".tmp")),
-			[],
-		);
-	});
 });
 
 describe("Anthropic persisted status and Pi refresh", () => {
